+++ src/machinable/component.py
@@ -102,6 +102,8 @@ class Component(Interface):
     def launch(self) -> Self:
         from machinable.execution import Execution
 
+        self.fetch()
+
         if Execution.is_connected():
             # commit only, defer execution
             Execution.get().add(self)